+++ milvus/client/grpc_handler.py
@@ -1,5 +1,4 @@
 import datetime
-from urllib.parse import urlparse
 import logging
 import threading
 import ujson
@@ -14,8 +13,6 @@ from .prepare import Prepare
 from .types import Status
 from .check import (
     int_or_str,
-    is_legal_host,
-    is_legal_port,
 )
 
 from .abs_client import AbsMilvus
@@ -23,9 +20,14 @@ from .asynch import SearchFuture, InsertFuture, CreateIndexFuture, CompactFuture
 
 from .hooks import BaseSearchHook
 from .client_hooks import SearchHook, HybridSearchHook
-from .exceptions import *
-from ..settings import DefaultConfig as config
-from . import __version__
+from .exceptions import (
+    BaseError,
+    CollectionNotExistException,
+    IllegalCollectionNameException,
+    NotConnectError,
+    ParamError
+)
+from .utils import set_uri
 
 LOGGER = logging.getLogger(__name__)
 
@@ -36,12 +38,12 @@ def error_handler(*rargs):
             record_dict = {}
             try:
                 record_dict["API start"] = str(datetime.datetime.now())
-                if self._pre_ping:
+                if getattr(self, '_pre_ping'):
                     self.ping()
                 record_dict["RPC start"] = str(datetime.datetime.now())
                 return func(self, *args, **kwargs)
             except BaseError as e:
-                LOGGER.error("Error: {}".format(e))
+                LOGGER.error("Error: {}", str(e))
                 if e.code == Status.ILLEGAL_COLLECTION_NAME:
                     raise IllegalCollectionNameException(e.code, e.message)
                 if e.code == Status.COLLECTION_NOT_EXISTS:
@@ -51,15 +53,18 @@ def error_handler(*rargs):
 
             except grpc.FutureTimeoutError as e:
                 record_dict["RPC timeout"] = str(datetime.datetime.now())
-                LOGGER.error("\nAddr [{}] {}\nRequest timeout: {}\n\t{}".format(self.server_address, func.__name__, e, record_dict))
+                LOGGER.error("\nAddr [{}] {}\nRequest timeout: {}\n\t{}",
+                             self.server_address, func.__name__, e, record_dict)
                 raise e
             except grpc.RpcError as e:
                 record_dict["RPC error"] = str(datetime.datetime.now())
-                LOGGER.error("\nAddr [{}] {}\nRPC error: {}\n\t{}".format(self.server_address, func.__name__, e, record_dict))
+                LOGGER.error("\nAddr [{}] {}\nRPC error: {}\n\t{}",
+                             self.server_address, func.__name__, e, record_dict)
                 raise e
             except Exception as e:
                 record_dict["Exception"] = str(datetime.datetime.now())
-                LOGGER.error("\nAddr [{}] {}\nExcepted error: {}\n\t{}".format(self.server_address, func.__name__, e, record_dict))
+                LOGGER.error("\nAddr [{}] {}\nExcepted error: {}\n\t{}",
+                             self.server_address, func.__name__, e, record_dict)
                 raise e
 
         return handler
@@ -67,33 +72,11 @@ def error_handler(*rargs):
     return wrapper
 
 
-def set_uri(host, port, uri):
-    if host is not None:
-        _port = port if port is not None else config.GRPC_PORT
-        _host = host
-    elif port is None:
-        try:
-            _uri = urlparse(uri) if uri else urlparse(config.GRPC_URI)
-            _host = _uri.hostname
-            _port = _uri.port
-        except (AttributeError, ValueError, TypeError) as e:
-            raise ParamError("uri is illegal: {}".format(e))
-    else:
-        raise ParamError("Param is not complete. Please invoke as follow:\n"
-                         "\t(host = ${HOST}, port = ${PORT})\n"
-                         "\t(uri = ${URI})\n")
-
-    if not is_legal_host(_host) or not is_legal_port(_port):
-        raise ParamError("host or port is illeagl")
-
-    return "{}:{}".format(str(_host), str(_port))
-
-
 class GrpcHandler(AbsMilvus):
-    def __init__(self, host=None, port=None, pre_ping=True, **kwargs):
+    def __init__(self, uri, pre_ping=True, **kwargs):
         self._channel = None
         self._stub = None
-        self._uri = None
+        self._uri = uri
         self.status = None
         self._connected = False
         self._pre_ping = pre_ping
@@ -113,8 +96,7 @@ class GrpcHandler(AbsMilvus):
         self._search_file_hook = SearchHook()
 
         # set server uri if object is initialized with parameter
-        _uri = kwargs.get("uri", None)
-        self._setup(host, port, _uri, pre_ping)
+        self._setup(uri, pre_ping)
 
     def __str__(self):
         attr_list = ['%s=%r' % (key, value)
@@ -127,14 +109,14 @@ class GrpcHandler(AbsMilvus):
     def __exit__(self, exc_type, exc_val, exc_tb):
         pass
 
-    def _setup(self, host, port, uri, pre_ping=False):
+    def _setup(self, uri, pre_ping=False):
         """
         Create a grpc channel and a stub
 
         :raises: NotConnectError
 
         """
-        self._uri = set_uri(host, port, uri)
+        self._uri = set_uri(uri)
         self._channel = grpc.insecure_channel(
             self._uri,
             options=[(cygrpc.ChannelArgKey.max_send_message_length, -1),
@@ -191,23 +173,23 @@ class GrpcHandler(AbsMilvus):
                 try:
                     ft.result(timeout=timeout)
                     return True
-                except:
+                except (grpc.RpcError, grpc.FutureTimeoutError):
                     retry -= 1
-                    LOGGER.debug("Retry connect addr <{}> {} times".format(self._uri, self._max_retry - retry))
+                    LOGGER.debug("Retry connect addr <{}> {} times",
+                                 self._uri, self._max_retry - retry)
                     if retry > 0:
                         timeout *= 2
                         continue
-                    else:
-                        LOGGER.error("Retry to connect server {} failed.".format(self._uri))
-                        raise
+
+                    LOGGER.error("Retry to connect server {} failed.", self._uri)
+                    raise
         except grpc.FutureTimeoutError:
-            raise NotConnectError('Fail connecting to server on {}. Timeout'.format(self._uri))
+            raise NotConnectError(f'Fail connecting to server on {self._uri}. Timeout')
         except grpc.RpcError as e:
-            raise NotConnectError("Connect error: <{}>".format(e))
+            raise NotConnectError(f"Connect error: <{e}>")
         # Unexpected error
         except Exception as e:
-            raise NotConnectError("Error occurred when trying to connect server:\n"
-                                  "\t<{}>".format(str(e)))
+            raise NotConnectError(f"Error occurred when trying to connect server:\n\t<{str(e)}>")
 
     @property
     def server_address(self):
@@ -243,7 +225,7 @@ class GrpcHandler(AbsMilvus):
         return self._cmd(cmd='status', timeout=timeout)
 
     @error_handler(None)
-    def _cmd(self, cmd, timeout=30):
+    def cmd(self, cmd, timeout=30):
         cmd = Prepare.cmd(cmd)
         rf = self._stub.Cmd.future(cmd, wait_for_ready=True, timeout=timeout)
         response = rf.result()
@@ -255,7 +237,9 @@ class GrpcHandler(AbsMilvus):
     @error_handler()
     def create_collection(self, collection_name, fields, timeout=30):
         collection_schema = Prepare.collection_schema(collection_name, fields)
-        rf = self._stub.CreateCollection.future(collection_schema, wait_for_ready=True, timeout=timeout)
+        rf = self._stub.CreateCollection.future(collection_schema,
+                                                wait_for_ready=True,
+                                                timeout=timeout)
         status = rf.result()
         if status.error_code != 0:
             LOGGER.error(status)
@@ -265,7 +249,9 @@ class GrpcHandler(AbsMilvus):
     def has_collection(self, collection_name, timeout=30, **kwargs):
         collection_name = Prepare.collection_name(collection_name)
 
-        rf = self._stub.HasCollection.future(collection_name, wait_for_ready=True, timeout=timeout)
+        rf = self._stub.HasCollection.future(collection_name,
+                                             wait_for_ready=True,
+                                             timeout=timeout)
         reply = rf.result()
         if reply.status.error_code == 0:
             return reply.bool_reply
@@ -275,7 +261,9 @@ class GrpcHandler(AbsMilvus):
     @error_handler(None)
     def get_collection_info(self, collection_name, timeout=30, **kwargs):
         collection_name = Prepare.collection_name(collection_name)
-        rf = self._stub.DescribeCollection.future(collection_name, wait_for_ready=True, timeout=timeout)
+        rf = self._stub.DescribeCollection.future(collection_name,
+                                                  wait_for_ready=True,
+                                                  timeout=timeout)
         response = rf.result()
 
         if response.status.error_code == 0:
@@ -288,7 +276,9 @@ class GrpcHandler(AbsMilvus):
     def count_entities(self, collection_name, timeout=30, **kwargs):
         collection_name = Prepare.collection_name(collection_name)
 
-        rf = self._stub.CountCollection.future(collection_name, wait_for_ready=True, timeout=timeout)
+        rf = self._stub.CountCollection.future(collection_name,
+                                               wait_for_ready=True,
+                                               timeout=timeout)
         response = rf.result()
         if response.status.error_code == 0:
             return response.collection_row_count
@@ -321,7 +311,9 @@ class GrpcHandler(AbsMilvus):
     @error_handler()
     def load_collection(self, collection_name, timeout=None):
         collection_name = Prepare.collection_name(collection_name)
-        status = self._stub.PreloadCollection.future(collection_name, wait_for_ready=True, timeout=timeout).result()
+        status = self._stub.PreloadCollection.future(collection_name,
+                                                     wait_for_ready=True,
+                                                     timeout=timeout).result()
         if status.error_code != 0:
             raise BaseError(status.error_code, status.reason)
 
@@ -329,7 +321,8 @@ class GrpcHandler(AbsMilvus):
     def reload_segments(self, collection_name, segment_ids, timeout=30):
         file_ids = list(map(int_or_str, segment_ids))
         request = Prepare.reload_param(collection_name, file_ids)
-        status = self._stub.ReloadSegments.future(request, wait_for_ready=True, timeout=timeout).result()
+        status = self._stub.ReloadSegments.future(request, wait_for_ready=True, timeout=timeout)\
+                                          .result()
         if status.error_code != 0:
             raise BaseError(status.error_code, status.reason)
 
@@ -337,7 +330,9 @@ class GrpcHandler(AbsMilvus):
     def drop_collection(self, collection_name, timeout=20):
         collection_name = Prepare.collection_name(collection_name)
 
-        rf = self._stub.DropCollection.future(collection_name, wait_for_ready=True, timeout=timeout)
+        rf = self._stub.DropCollection.future(collection_name,
+                                              wait_for_ready=True,
+                                              timeout=timeout)
         status = rf.result()
         if status.error_code != 0:
             raise BaseError(status.error_code, status.reason)
@@ -345,7 +340,8 @@ class GrpcHandler(AbsMilvus):
         return Status(status.error_code, status.reason)
 
     @error_handler([])
-    def bulk_insert(self, collection_name, entities, types, ids=None, partition_tag=None, params=None, timeout=None, **kwargs):
+    def bulk_insert(self, collection_name, entities, types, ids=None,
+                    partition_tag=None, params=None, timeout=None, **kwargs):
         insert_param = kwargs.get('insert_param', None)
 
         if insert_param and not isinstance(insert_param, grpc_types.InsertParam):
@@ -381,7 +377,8 @@ class GrpcHandler(AbsMilvus):
 
     @error_handler([])
     def list_id_in_segment(self, collection_name, segment_id, timeout=30):
-        request = grpc_types.GetEntityIDsParam(collection_name=collection_name, segment_id=segment_id)
+        request = grpc_types.GetEntityIDsParam(collection_name=collection_name,
+                                               segment_id=segment_id)
 
         rf = self._stub.GetEntityIDs.future(request, wait_for_ready=True, timeout=timeout)
         response = rf.result()
@@ -443,7 +440,7 @@ class GrpcHandler(AbsMilvus):
         response = rf.result()
         status = response.status
         if status.error_code == 0:
-            return [p for p in response.partition_tag_array]
+            return list(response.partition_tag_array)
 
         raise BaseError(status.error_code, status.reason)
 
@@ -460,8 +457,8 @@ class GrpcHandler(AbsMilvus):
         return Status(response.error_code, response.reason)
 
     @error_handler(None)
-    def search(self, collection_name, query_entities, partition_tags=None, fields=None, **kwargs):
-        request = Prepare.search_param(collection_name, query_entities, partition_tags, fields)
+    def search(self, collection_name, dsl, partition_tags=None, fields=None, **kwargs):
+        request = Prepare.search_param(collection_name, dsl, partition_tags, fields)
         self._search_hook.pre_search()
         to = kwargs.get("timeout", None)
         ft = self._stub.Search.future(request, wait_for_ready=True, timeout=to)
@@ -478,13 +475,11 @@ class GrpcHandler(AbsMilvus):
         if response.status.error_code != 0:
             raise BaseError(response.status.error_code, response.status.reason)
 
-        # TODO: handler response
-        # resutls = self._search_hook.handle_response(response)
-
         return QueryResult(response)
 
     @error_handler(None)
-    def search_by_ids(self, collection_name, ids, top_k, partition_tags=None, params=None, timeout=None, **kwargs):
+    def search_by_ids(self, collection_name, ids, top_k, partition_tags=None,
+                      params=None, timeout=None, **kwargs):
         request = Prepare.search_by_ids_param(collection_name, ids, top_k, partition_tags, params)
         if kwargs.get("_async", False) is True:
             future = self._stub.SearchByID.future(request, wait_for_ready=True, timeout=timeout)
@@ -551,7 +546,7 @@ class GrpcHandler(AbsMilvus):
         if kwargs.get("_async", False):
             cb = kwargs.get("_callback", None)
             return FlushFuture(future, cb)
-        response = future.result()
+        response = future.done()
         if response.error_code != 0:
             raise BaseError(response.error_code, response.reason)