+++ milvus/client/stub.py
@@ -11,7 +11,6 @@ from collections import defaultdict
 from urllib.parse import urlparse
 
 from . import __version__
-from .types import Status
 from .check import check_pass_param, is_legal_host, is_legal_port
 from .pool import ConnectionPool, SingleConnectionPool, SingletonThreadPool
 from .exceptions import ParamError, DeprecatedError
@@ -95,6 +94,9 @@ class Milvus:
         else:
             raise ParamError("Unknown pool value: {}".format(pool))
 
+        #
+        self._conn = None
+
         # store extra key-words arguments
         self._kw = kwargs
         self._hooks = collections.defaultdict()
@@ -173,7 +175,7 @@ class Milvus:
         check_pass_param(cmd=cmd)
 
         with self._connection() as handler:
-            return handler._cmd(cmd, timeout)
+            return handler.cmd(cmd, timeout)
 
     @check_connect
     def create_collection(self, collection_name, fields, timeout=30):
@@ -279,7 +281,7 @@ class Milvus:
 
         :param collection_name: The name of the collection to get statistics about.
         :type  collection_name: str
-        
+
         :return: The collection stats.
         :rtype: dict
 
@@ -327,7 +329,8 @@ class Milvus:
             return handler.drop_collection(collection_name, timeout)
 
     @check_connect
-    def bulk_insert(self, collection_name, entities, ids=None, partition_tag=None, params=None, timeout=None, **kwargs):
+    def bulk_insert(self, collection_name, entities, ids=None,
+                    partition_tag=None, params=None, timeout=None, **kwargs):
         """
         Inserts columnar entities in a specified collection.
 
@@ -368,7 +371,8 @@ class Milvus:
         if ids is not None:
             check_pass_param(ids=ids)
         with self._connection() as handler:
-            results = handler.bulk_insert(collection_name, entities, copy_fields, ids, partition_tag, params, timeout, **kwargs)
+            results = handler.bulk_insert(collection_name, entities, copy_fields,
+                                          ids, partition_tag, params, timeout, **kwargs)
             with self._cache_cv:
                 self._c_cache[collection_name] = copy_fields
             return results
@@ -403,7 +407,8 @@ class Milvus:
 
         :param collection_name: The name of the collection that contains the specified segment
         :type  collection_name: str
-        :param segment_id: The ID of the segment. You can get segment IDs by calling the get_collection_stats() method.
+        :param segment_id: The ID of the segment. You can get segment IDs by calling the
+                           get_collection_stats() method.
         :type  segment_id: int
 
         :return: List of IDs in a specified segment.
@@ -422,10 +427,11 @@ class Milvus:
     @check_connect
     def create_index(self, collection_name, field_name, params, timeout=None, **kwargs):
         """
-        Creates an index for a field in a specified collection. Milvus does not support creating multiple
-        indexes for a field. In a scenario where the field already has an index, if you create another one
-        that is equivalent (in terms of type and parameters) to the existing one, the server returns this
-        index to the client; otherwise, the server replaces the existing index with the new one.
+        Creates an index for a field in a specified collection. Milvus does not support
+        creating multiple indexes for a field. In a scenario where the field already has
+        an index, if you create another one that is equivalent (in terms of type and
+        parameters) to the existing one, the server returns this index to the client;
+        otherwise, the server replaces the existing index with the new one.
 
         :param collection_name: The name of the collection to create field indexes.
         :type  collection_name: str
@@ -555,7 +561,8 @@ class Milvus:
             return handler.drop_partition(collection_name, partition_tag, timeout)
 
     @check_connect
-    def search(self, collection_name, dsl, partition_tags=None, fields=None, timeout=None, **kwargs):
+    def search(self, collection_name, dsl, partition_tags=None,
+               fields=None, timeout=None, **kwargs):
         """
         Searches a collection based on the given DSL clauses and returns query results.
 
@@ -577,16 +584,18 @@ class Milvus:
             BaseError: If the return result from server is not ok
         """
         with self._connection() as handler:
-            return handler.search(collection_name, dsl, partition_tags, fields, timeout=timeout, **kwargs)
+            return handler.search(collection_name, dsl, partition_tags,
+                                  fields, timeout=timeout, **kwargs)
 
     @check_connect
-    def search_in_segment(self, collection_name, segment_ids, dsl, fields=None, timeout=None, **kwargs):
+    def search_in_segment(self, collection_name, segment_ids, dsl,
+                          fields=None, timeout=None, **kwargs):
         """
         Searches in the specified segments of a collection.
 
-        The Milvus server stores entity data into multiple files. Searching for entities in specific files is a
-        method used in Mishards. Obtain more detail about Mishards, see
-        <a href="https://github.com/milvus-io/milvus/tree/master/shards">
+        The Milvus server stores entity data into multiple files. Searching for entities
+        in specific files is a method used in Mishards. Obtain more detail about Mishards,
+        see <a href="https://github.com/milvus-io/milvus/tree/master/shards">
 
         :param collection_name: The name of the collection to search.
         :type  collection_name: str:param collection_name: table name been queried
@@ -605,13 +614,13 @@ class Milvus:
             ParamError: If parameters are invalid
             BaseError: If the return result from server is not ok
         """
-        # check_pass_param(collection_name=collection_name, segment_ids, query_entities, params, timeout)
 
         # params = dict() if params is None else params
         # if not isinstance(params, dict):
         #     raise ParamError("Params must be a dictionary type")
         with self._connection() as handler:
-            return handler.search_in_segment(collection_name, segment_ids, dsl, fields, timeout, **kwargs)
+            return handler.search_in_segment(collection_name, segment_ids, dsl,
+                                             fields, timeout, **kwargs)
 
     @check_connect
     def delete_entity_by_id(self, collection_name, ids, timeout=None):
@@ -640,10 +649,10 @@ class Milvus:
     @check_connect
     def flush(self, collection_name_array=None, timeout=None, **kwargs):
         """
-        Flushes data in the specified collections from memory to disk. When you insert or delete data,
-        the server stores the data in the memory temporarily and then flushes it to the disk at fixed
-        intervals. Calling flush ensures that the newly inserted data is visible and the deleted data
-        is no longer recoverable.
+        Flushes data in the specified collections from memory to disk. When you insert or
+        delete data, the server stores the data in the memory temporarily and then flushes
+        it to the disk at fixed intervals. Calling flush ensures that the newly inserted
+        data is visible and the deleted data is no longer recoverable.
 
         :type  collection_name_array: An array of names of the collections to flush.
         :param collection_name_array: list[str]
@@ -714,4 +723,3 @@ class Milvus:
         cmd = "SET {} {}".format(key, value)
 
         return self._cmd(cmd)
-